HttpClient Enhancements (#60)
* HttpClient Enhancements

The 2 enhancements below have been developed to address the concerns raised with the HttpClient being created/disposed on every request, and requests for the ability to have access to the HttpClient itself from the consuming application. (detailed in #57 and #58).

1. Addressing the HttpClient being created/disposed on every request:
The HttpClient is not a thread-safe in entirety and there are other considerations that were required to ensure that the issue addressed all concerns. Utilizing the HttpClient as a singleton throughout the application can introduce issues in high load environments such as deadlocks, potential memory leaks, and potentially affect fail-over and load balancing scenarios where underlying socket sessions will prevent host switching.

To overcome this, we added a new LimitedPool and LimitedPoolItem class and refactored them to suit the requirements of SDK. These refactored classes do not use the HttpClient as a singleton which ensures thead-safe and deadlocks are not hit, and are configurable to specify the lifetime of the cached HttpClient objects which avoids potential fail-over and load balancing issues.

To specify a specific lifetime of the cached HttpClient objects, you either supply the optional constructor parameter httpClientPoolItemLifetime to the Configuration class, or you set it directly in the Configuration class through the HttpClientPoolItemLifetime property.

When the Configuration class HttpClientPoolItemLifetime property has not been specifically set via the constructor or direct property access, the HttpClientPoolItem and its associated HttpClient will have a lifetime of 1 hour.

2. Providing the ability to have access to the HttpClient itself from the consuming application via a constructor and/or public property. NOTE: When utilizing it this way, you will in effect bypass the pooling entirely and must supply the HttpClient that is to be used.
Using this method of HttpClient consumption, you supply the Client class with the HttpClient to be utilized – whether it be during instantiation through the optional httpClientInstance constructor parameter, or afterwards through the HttpClientInstance property accessor.

Overall Refactoring Specifics:
- Added optional httpClientPoolItemLifetime parameter to constructor of Configuration class. If unspecified, will utilize a default value of 1 hour.
- Added public get/set HttpClientPoolItemLifetime property to Configuration class.
- Added optional httpClientInstance parameter to constructor of Client class.
- Added public get/set HttpClientInstance property to Client class.
- Added private static get access to new HttpClientPool property with instantiation lock.
- Added and refactored LimitedPool and LimitedPoolItem class for SDK specific usage with IDisposable objects and varied configuration lifetimes.
- Refactored all HttpClient accesses in Client to utilize the HttpClientInstance if not null, otherwise it will draw an instance from the HttpClientPool.

* introduced IClient and update all managers and extension methods to use the Iclient instead of Client object.

- Added IClient interface to be able to refactor the client implementation